Create resources.md
Still need to continue adding more resources to the list, and go back to find the hyperlinks for them. Would it be more helpful to link each citation to its corresponding document in the Google Drive folder I created, or would it be better to link it to the database in which I found it?